+++ SocialBehaviorptc/ssm_ptc/transitions/stationary_transition.py
@@ -8,7 +8,7 @@ from ssm_ptc.utils import set_param
 
 class StationaryTransition(BaseTransition):
 
-    def __init__(self, K, D, M=0, Pi=None, device=torch.device('cpu'), **kwargs):
+    def __init__(self, K, D, M=0, Pi=None, device=torch.device('cpu'), dtype=torch.float64, **kwargs):
         super(StationaryTransition, self).__init__(K, D, M, device=device)
 
         if Pi is None:
@@ -17,7 +17,7 @@ class StationaryTransition(BaseTransition):
             assert isinstance(Pi, np.ndarray)
             assert Pi.shape == (K, K)
 
-        self.Pi = torch.tensor(Pi, dtype=torch.float64, device=device, requires_grad=True)
+        self.Pi = torch.tensor(Pi, dtype=dtype, device=device, requires_grad=True)
 
     @property
     def params(self):
